% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_datasets}
\alias{check_datasets}
\title{Check a list of sampled datasets for non-varying variables. A mclapply wrapper around check_varying()}
\usage{
check_datasets(dt, samples, x, y, type = "lm")
}
\arguments{
\item{dt}{the initial dataset}

\item{samples}{which rows to subset from dt. A list of list, with at least a samples$rows containing the rows indices to subset. samples$rows can be greater than nrow(dt).}

\item{x}{character vector, the x variable}

\item{y}{character vector, the y variable}
}
\value{
The samples list with added samples$use = FALSE if the sampled dt should not be used, samples$use = TRUE otherwise. If not null, a samples$to_remove character vector of non-varying variables to remove from the model for the given sample.
}
\description{
Check a list of sampled datasets for non-varying variables. A mclapply wrapper around check_varying()
}
