% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_varying}
\alias{check_varying}
\title{Function that checks for non-varying variables}
\usage{
check_varying(data, x, y, type = "lm")
}
\arguments{
\item{data}{the dataset}

\item{x}{the name of the x variable of the model}

\item{y}{the name of the y variable of the model}
}
\value{
returns TRUE if all variables vary. Returns FALSE if x or y vary. Else returns the names of the non-varying variables.
}
\description{
Function that checks for non-varying variables
}
