% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_inclusion}
\alias{compute_inclusion}
\title{Compute proportion of covar inclusion in the model}
\usage{
compute_inclusion(res, covars, x, y, size)
}
\arguments{
\item{res}{the list of models}

\item{covars}{character of covariables names}

\item{x}{character vector, the x variable}

\item{y}{character vector, the y variable}

\item{samples}{the list returned by check_datasets()}
}
\description{
Compute proportion of covar inclusion in the model
}
