% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_covar}
\alias{get_covar}
\title{Get covariables names included in the model}
\usage{
get_covar(model, x, y)
}
\arguments{
\item{model}{the model}

\item{x}{character vector, the x variable}

\item{y}{character vector, the y variable}
}
\description{
Get covariables names included in the model
}
