% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{make_samples}
\alias{make_samples}
\title{Make and check samples from the intial dataset}
\usage{
make_samples(params, sizes, N, cores)
}
\arguments{
\item{sizes}{a named numeric vector of the samples sizes}

\item{N}{an integer, the number of samples to make}

\item{cores}{integer, the number of cores to use for parallel computing}

\item{dt}{the dataset}

\item{y}{}
}
\value{
a named list of samples s, with s$rows the rows to take from the initial dataset. s$use whether the sample dataset can be used. s$to_remove non-varying variables to remove.
}
\description{
Make and check samples from the intial dataset
}
