% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_full_model}
\alias{run_full_model}
\title{Function to run a saturated linear model}
\usage{
run_full_model(data, x, y, type = "lm")
}
\arguments{
\item{data}{the data set}

\item{x}{character vector, the x variable}

\item{y}{character vector, the y variable}
}
\value{
the linear model object
}
\description{
Function to run a saturated linear model
}
